Add new properties to QuickTaskInfo and comment out instantiation
Two new properties, `IsEmpty` and `IsUploadTask`, have been added to the `QuickTaskInfo` class. `IsEmpty` checks if the `AfterCaptureTasks` property is set to `None`, while `IsUploadTask` checks if the `AfterCaptureTasks` property has the `UploadImageToHost` flag set. Additionally, the instantiation of `QuickTaskInfo` with "Annotate" as the `Name` and `AfterCaptureTasks.AnnotateImage` as the `AfterCaptureTasks` has been commented out in the array of `QuickTaskInfo` objects.